sms: Combine SMS_tx and rx into single node
