+++ ogc_ws/src/sms/launch/sms_standalone.launch
@@ -2,8 +2,9 @@
 
 <launch>
 
-    <!-- please fill in the ground router's number. Enclose it in '' to recognise it as a string (instead of int)-->
-    <arg name="GCS_no" default="'12345678'"/>
+    <!-- please fill in the phone no of the client (e.g. if running this on air, fill in ground phone no,
+     and vice versa). Enclose it in '' to recognise it as a string (instead of int)-->
+    <arg name="client_phone_no" default="'12345678'"/>
 
     <!-- phone numbers authorized to send commands to aircraft. Whitelist should be in same folder as nodes-->
     <arg name="whitelist" default="$(find sms)/src/whitelist.txt"/>
@@ -18,23 +19,15 @@
     <!-- hostname and IP of onboard cellular router -->
     <arg name="router_hostname" default="root@192.168.1.1" />
 
-    <!-- SMS_rx node: Receive SMS from Ground Control and send commands to aircraft/SMS_tx if necessary-->
-    <node pkg="sms" type="SMS_rx.py" name="SMS_rx"
+    <!-- sms_link node -->
+    <node pkg="sms" type="sms_link.py" name="sms_link"
     respawn="true" respawn_delay="3" output="$(arg log_output)">
         <param name="whitelist" value="$(arg whitelist)" />
         <param name="router_hostname" value = "$(arg router_hostname)"/>
+        <param name="client_phone_no" value="$(arg client_phone_no)" />
     </node>
 
-    <!--SMS_tx node: Send SMS containing key aircraft info to Ground Control.
-    SMS sending is deactivated by default. If SMS_rx receives a message by GCS requesting for SMS sending,
-    then SMS_rx will notify SMS_tx to start sending (similar for deactivation)-->
-    <node pkg="sms" type="SMS_tx.py" name="SMS_tx"
-    respawn="true" respawn_delay="3" output="$(arg log_output)">
-        <param name="GCS_no" value="$(arg GCS_no)" />
-        <param name="router_hostname" value = "$(arg router_hostname)"/>
-    </node>
-
-        <!--air_despatcher node-->
+    <!-- air_despatcher node -->
     <node pkg="despatcher" type="air_despatcher.py" name="air_despatcher"
     respawn="true" respawn_delay="3" output="$(arg log_output)">
         <param name="interval_1" value="$(arg interval_1)" />
